app/requirements.txt
fastapi
selenium
pytesseract
pdf2image
psycopg2-binary
uvicorn
ormar[postgresql]
eventkit
pydantic-settings

